// src/schemas/notifications-api.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
	"/updates": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/** Subscribe to notification updates */
		get: {
			parameters: {
				query?: never;
				header?: never;
				path?: never;
				cookie?: never;
			};
			requestBody?: never;
			responses: {
				/** @description OK */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"text/event_stream":
							| components["schemas"]["EventNotifications"]
							| components["schemas"]["EventNotificationUpdate"];
					};
				};
			};
		};
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
}
export type webhooks = Record<string, never>;
export interface components {
	schemas: {
		Notification: {
			/** @example 1337 */
			id?: number;
			/** @example nics.unplug.ended for compute/instances/27000000004387 */
			summary?: string;
			/** @example Network Interface nic1 (VirtIO) was unplugged from VM devel_dz-shy-dew-01-764. (User: admin@internal-authz) */
			body?: string;
			/** @example compute */
			service?: string;
			/** @example Info */
			severity?: string;
			/** @example 2025-04-09T11:21:21Z */
			event_at?: string;
			/** @example 7b96e5fe-753b-4981-9fef-0c20bf833c55 */
			event_id?: string;
			/** @example 7 */
			user_id?: number;
			/** @example 887 */
			namespace_id?: number;
			/** @example 2025-04-09T11:21:21Z */
			read_at?: string;
			/** @example 2025-04-09T11:21:21Z */
			created_at?: string;
			/** @example 2025-04-09T11:21:21Z */
			updated_at?: string;
		};
		EventNotifications: {
			/** @example notifications */
			event?: string;
			data?: components["schemas"]["Notification"][];
		};
		EventNotificationUpdate: {
			/** @example notification_update */
			event?: string;
			data?: components["schemas"]["Notification"][];
		};
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;